Translation-Evaluation/requirements.txt
transformers>=4.40.0
torch>=2.2.0
sacrebleu>=2.4.0
unbabel-comet
google-cloud-translate
nltk>=3.8.1
hf_xet
dotenv
datasets